+++ kamu-core/src/infra/ingest_service_impl.rs
@@ -34,47 +34,6 @@ impl IngestServiceImpl {
     fn get_dataset_layout(&self, dataset_id: &DatasetID) -> DatasetLayout {
         DatasetLayout::create(&self.volume_layout, dataset_id).unwrap()
     }
-
-    fn update_summary(
-        &self,
-        dataset_id: &DatasetID,
-        result: &IngestResult,
-    ) -> Result<(), IngestError> {
-        match result {
-            IngestResult::UpToDate => Ok(()),
-            IngestResult::Updated {
-                block_hash,
-                has_more: _,
-            } => {
-                let mut metadata_repo = self.metadata_repo.borrow_mut();
-
-                let mut summary = metadata_repo
-                    .get_summary(dataset_id)
-                    .map_err(|e| IngestError::internal(e))?;
-
-                let block = metadata_repo
-                    .get_metadata_chain(dataset_id)
-                    .unwrap()
-                    .get_block(block_hash)
-                    .unwrap();
-
-                summary.num_records = match block.output_slice {
-                    Some(slice) => summary.num_records + slice.num_records as u64,
-                    _ => 0,
-                };
-
-                summary.last_pulled = Some(block.system_time);
-
-                let layout = DatasetLayout::new(&self.volume_layout, dataset_id);
-                summary.data_size = fs_extra::dir::get_size(layout.data_dir).unwrap_or(0);
-                summary.data_size += fs_extra::dir::get_size(layout.checkpoints_dir).unwrap_or(0);
-
-                metadata_repo
-                    .update_summary(dataset_id, summary)
-                    .map_err(|e| IngestError::internal(e))
-            }
-        }
-    }
 }
 
 impl IngestService for IngestServiceImpl {
@@ -110,9 +69,7 @@ impl IngestService for IngestServiceImpl {
             logger,
         );
 
-        let result = ingest_task.ingest()?;
-        self.update_summary(dataset_id, &result)?;
-        Ok(result)
+        ingest_task.ingest()
     }
 
     fn ingest_multi(
@@ -184,14 +141,6 @@ impl IngestService for IngestServiceImpl {
             .into_iter()
             .flat_map(|h| h.join().unwrap())
             .collect();
-
-        results
-            .iter()
-            .filter(|(_, res)| res.is_ok())
-            .for_each(|(dataset_id, res)| {
-                self.update_summary(dataset_id, res.as_ref().unwrap())
-                    .unwrap()
-            });
 
         results
     }