+++ kamu-core/src/infra/metadata_repository_impl.rs
@@ -6,7 +6,7 @@ use chrono::Utc;
 use std::collections::HashSet;
 use std::collections::LinkedList;
 use std::convert::TryFrom;
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
 use url::Url;
 
 pub struct MetadataRepositoryImpl {
@@ -39,7 +39,7 @@ impl MetadataRepositoryImpl {
         &self,
         dataset_id: &DatasetID,
     ) -> Result<MetadataChainImpl, DomainError> {
-        let path = self.workspace_layout.datasets_dir.join(dataset_id.as_str());
+        let path = self.get_dataset_metadata_dir(dataset_id);
         if !path.exists() {
             Err(DomainError::does_not_exist(
                 ResourceKind::Dataset,
@@ -77,6 +77,80 @@ impl MetadataRepositoryImpl {
         }
         ordered
     }
+
+    fn read_summary(&self, path: &Path) -> Result<DatasetSummary, DomainError> {
+        let file = std::fs::File::open(&path).unwrap_or_else(|e| {
+            panic!(
+                "Failed to open the summary file at {}: {}",
+                path.display(),
+                e
+            )
+        });
+
+        let manifest: Manifest<DatasetSummary> =
+            serde_yaml::from_reader(&file).unwrap_or_else(|e| {
+                panic!(
+                    "Failed to deserialize the DatasetSummary at {}: {}",
+                    path.display(),
+                    e
+                )
+            });
+
+        assert_eq!(manifest.kind, "DatasetSummary");
+        Ok(manifest.content)
+    }
+
+    fn write_summary(&self, path: &Path, summary: &DatasetSummary) -> Result<(), DomainError> {
+        let manifest = Manifest {
+            api_version: 1,
+            kind: "DatasetSummary".to_owned(),
+            content: summary.clone(),
+        };
+        let file = std::fs::File::create(&path).map_err(|e| InfraError::from(e).into())?;
+        serde_yaml::to_writer(file, &manifest).map_err(|e| InfraError::from(e).into())?;
+        Ok(())
+    }
+
+    // TODO: summaries should be per branch
+    fn summarize(&self, dataset_id: &DatasetID) -> Result<DatasetSummary, DomainError> {
+        let chain = self.get_metadata_chain_impl(dataset_id)?;
+
+        let mut summary = DatasetSummary {
+            id: dataset_id.to_owned(),
+            last_block_hash: chain.read_ref(&BlockRef::Head).unwrap(),
+            ..DatasetSummary::default()
+        };
+
+        for block in chain.iter_blocks() {
+            match block.source {
+                Some(DatasetSource::Root(_)) => {
+                    summary.kind = DatasetKind::Root;
+                }
+                Some(DatasetSource::Derivative(src)) => {
+                    summary.kind = DatasetKind::Derivative;
+                    if summary.dependencies.is_empty() {
+                        summary.dependencies = src.inputs;
+                    }
+                }
+                None => (),
+            }
+
+            if let Some(slice) = block.output_slice {
+                if summary.last_pulled.is_none() {
+                    summary.last_pulled = Some(block.system_time);
+                }
+                summary.num_records += slice.num_records as u64;
+            }
+        }
+
+        // Calculate data size
+        let volume_layout = VolumeLayout::new(&self.workspace_layout.local_volume_dir);
+        let layout = DatasetLayout::new(&volume_layout, dataset_id);
+        summary.data_size = fs_extra::dir::get_size(layout.data_dir).unwrap_or(0);
+        summary.data_size += fs_extra::dir::get_size(layout.checkpoints_dir).unwrap_or(0);
+
+        Ok(summary)
+    }
 }
 
 impl MetadataRepository for MetadataRepositoryImpl {
@@ -94,21 +168,17 @@ impl MetadataRepository for MetadataRepositoryImpl {
             ));
         }
 
-        let (kind, dependencies) = match snapshot.source {
-            DatasetSource::Derivative(ref src) => {
-                for input_id in src.inputs.iter() {
-                    if !self.dataset_exists(input_id) {
-                        return Err(DomainError::missing_reference(
-                            ResourceKind::Dataset,
-                            String::from(&snapshot.id as &str),
-                            ResourceKind::Dataset,
-                            String::from(input_id as &str),
-                        ));
-                    }
+        if let DatasetSource::Derivative(ref src) = snapshot.source {
+            for input_id in src.inputs.iter() {
+                if !self.dataset_exists(input_id) {
+                    return Err(DomainError::missing_reference(
+                        ResourceKind::Dataset,
+                        String::from(&snapshot.id as &str),
+                        ResourceKind::Dataset,
+                        String::from(input_id as &str),
+                    ));
                 }
-                (DatasetKind::Derivative, src.inputs.clone())
             }
-            DatasetSource::Root { .. } => (DatasetKind::Root, Vec::new()),
         };
 
         let first_block = MetadataBlock {
@@ -125,16 +195,6 @@ impl MetadataRepository for MetadataRepositoryImpl {
         let (_chain, block_hash) =
             MetadataChainImpl::create(&dataset_metadata_dir, first_block).map_err(|e| e.into())?;
 
-        let summary = DatasetSummary {
-            id: snapshot.id.clone(),
-            kind: kind,
-            dependencies: dependencies,
-            last_pulled: None,
-            num_records: 0,
-            data_size: 0,
-        };
-
-        self.update_summary(&snapshot.id, summary)?;
         Ok(block_hash)
     }
 
@@ -217,62 +277,31 @@ impl MetadataRepository for MetadataRepositoryImpl {
     }
 
     fn get_summary(&self, dataset_id: &DatasetID) -> Result<DatasetSummary, DomainError> {
-        let path = self
-            .workspace_layout
-            .datasets_dir
-            .join(dataset_id)
-            .join("summary");
-        if !path.exists() {
-            Err(DomainError::does_not_exist(
+        if !self.dataset_exists(dataset_id) {
+            return Err(DomainError::does_not_exist(
                 ResourceKind::Dataset,
                 dataset_id.as_str().to_owned(),
-            ))
-        } else {
-            let file = std::fs::File::open(&path).unwrap_or_else(|e| {
-                panic!(
-                    "Failed to open the summary file at {}: {}",
-                    path.display(),
-                    e
-                )
-            });
-
-            let manifest: Manifest<DatasetSummary> =
-                serde_yaml::from_reader(&file).unwrap_or_else(|e| {
-                    panic!(
-                        "Failed to deserialize the DatasetSummary at {}: {}",
-                        path.display(),
-                        e
-                    )
-                });
-
-            assert_eq!(manifest.kind, "DatasetSummary");
-            Ok(manifest.content)
+            ));
         }
-    }
 
-    // TODO: summaries should be per branch
-    // TODO: vocab should be stored in the chain
-    // TODO: update summary lazily when new blocks appear
-    fn update_summary(
-        &mut self,
-        dataset_id: &DatasetID,
-        summary: DatasetSummary,
-    ) -> Result<(), DomainError> {
-        let path = self
-            .workspace_layout
-            .datasets_dir
-            .join(dataset_id)
-            .join("summary");
+        let path = self.get_dataset_metadata_dir(dataset_id).join("summary");
 
-        let manifest = Manifest {
-            api_version: 1,
-            kind: "DatasetSummary".to_owned(),
-            content: summary,
+        let summary = if path.exists() {
+            self.read_summary(&path)?
+        } else {
+            DatasetSummary::default()
         };
 
-        let file = std::fs::File::create(&path).map_err(|e| InfraError::from(e).into())?;
-        serde_yaml::to_writer(file, &manifest).map_err(|e| InfraError::from(e).into())?;
-        Ok(())
+        let chain = self.get_metadata_chain_impl(dataset_id)?;
+        let last_block_hash = chain.read_ref(&BlockRef::Head).unwrap();
+
+        if last_block_hash != summary.last_block_hash {
+            let summary = self.summarize(dataset_id)?;
+            self.write_summary(&path, &summary)?;
+            Ok(summary)
+        } else {
+            Ok(summary)
+        }
     }
 
     fn get_all_remotes<'s>(&'s self) -> Box<dyn Iterator<Item = RemoteIDBuf> + 's> {