+++ CHANGELOG.md
@@ -4,9 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.37.0] - 2020-12-27
+## Unreleased
 ### Changed
-- Breaking: Metadata restructuring
+- Breaking: Metadata restructuring means you'll need to re-create your datasets
+- Vocabulary has been incorporated into the metadata block (ODF `0.17.0`)
+- Lazily computing dataset summaries
+- Upgraded many dependencies
 
 ## [0.36.0] - 2020-11-16
 ### Changed